requirements.txt
python-binance==1.0.19
websockets>=12.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.1.0
colorlog>=6.8.0
asyncio-throttle>=1.0.2
pydantic>=2.5.0
pydantic-settings>=2.1.0
rich>=13.7.0
uvloop>=0.19.0; sys_platform != "win32"
orjson>=3.9.0
